+++ demos/src/b2c-palette.mustache
@@ -1,10 +1,10 @@
-	<div class="demo-wrapper">
-		{{#b2c}}
-			<div class="demo-color">
-				<h4 class="o-colors-border-{{name}}">{{name}}</h4>
-				<div class="swatch o-colors-palette-{{name}}" data-o-color="{{name}}">
-					<input type="text" class="hex" id="o-color-input-{{name}}" value="" readonly>
-				</div>
+<div class="demo-wrapper">
+	{{#b2c}}
+		<div class="demo-color">
+			<h4 class="o-colors-border-{{name}}">{{name}}</h4>
+			<div class="swatch o-colors-palette-{{name}}" data-o-color="{{name}}">
+				<input type="text" class="hex" id="o-color-input-{{name}}" value="" readonly>
 			</div>
-		{{/b2c}}
-	</div>
+		</div>
+	{{/b2c}}
+</div>
