+++ demos/src/color-mixer/color-mixer.scss
