+++ demos/src/contrast-checker/fieldsets/background.mustache
@@ -1,7 +1,7 @@
 <span class="o-forms-title__main">Palette</span>
 {{#primary}}
 	<label title="{{name}}">
-		<input type=radio name="background" style="background-color: var(--o-colors-{{name}})" value="{{name}}"  {{#baseBackground}}checked{{/baseBackground}}>
+		<input type=radio name="background" style="background-color: var(--o-colors-{{name}})" value="{{name}}" {{#baseBackground}}checked{{/baseBackground}}>
 	</label>
 {{/primary}}
 <br>
