+++ demos/src/contrast-checker/fieldsets/foreground.mustache
@@ -1,7 +1,7 @@
 <span class="o-forms-title__main">Palette</span>
 {{#primary}}
 	<label title="{{name}}">
-		<input type=radio name="foreground" style="background-color: var(--o-colors-{{name}})" value="{{name}}"  {{#baseText}}checked{{/baseText}}>
+		<input type=radio name="foreground" style="background-color: var(--o-colors-{{name}})" value="{{name}}" {{#baseText}}checked{{/baseText}}>
 	</label>
 {{/primary}}
 <br>
