--- demos/src/sandboxes/colors-mix.scss
@@ -1,82 +1,0 @@
-$o-colors-is-silent: false;
-
-// Import colors module
-@import '../../../main';
-
-:root {
-	--color: black;
-}
-
-.container {
-	max-width: 800px;
-	margin: 0 auto;
-	padding: 20px;
-	display: grid;
-	grid-template-columns: repeat(2, 0.75fr) 1fr;
-
-	.o-forms__label,
-	.o-forms__additional-info {
-		color: inherit;
-	}
-
-
-	@media (max-width: 630px) {
-		grid-template-columns: 1fr;
-
-		.details {
-			grid-row: 2;
-			padding-bottom: 20px;
-		}
-	}
-}
-
-.color-selectors {
-	padding: 20px;
-}
-
-body {
-	font-family: MetricWeb, sans-serif;
-	color: var(--color);
-}
-
-
-.details  {
-	pre {
-		margin: 10px 0;
-		padding: 10px;
-		border: 1px solid oColorsMix('black', 'white', 5);
-	}
-}
-
-.mix-range {
-	display: flex;
-	flex-direction: column-reverse;
-	justify-self: center;
-
-	[class^='percent'] {
-		display: flex;
-		justify-content: flex-start;
-	}
-
-	.sqr {
-		height: 2em;
-		width: 7em;
-		border: 1px solid var(--color);
-		margin: 2px;
-
-		&:hover {
-			cursor: pointer;
-			outline: 2px solid var(--color);
-		}
-
-		&[aria-selected=true] {
-			outline: 2px solid var(--color);
-		}
-	}
-
-	.text {
-		align-self: center;
-		padding-left: 0.5em;
-		font-weight: 600;
-	}
-}
