+++ origami.json
@@ -36,20 +36,20 @@
 			"name": "master-contrast-checker",
 			"title": "Contrast ratio checker",
 			"data": "demos/src/palettes/master-palette.json",
-			"js": "demos/src/sandboxes/contrast-checker.js",
-			"sass": "demos/src/sandboxes/contrast-checker.scss",
-			"template": "demos/src/sandboxes/contrast-checker.mustache",
+			"js": "demos/src/contrast-checker/contrast-checker.js",
+			"sass": "demos/src/contrast-checker/contrast-checker.scss",
+			"template": "demos/src/contrast-checker/index.mustache",
 			"display_html": false,
 			"brands": [
 				"master"
 			]
 		},
 		{
-			"name": "master-colors-mix",
+			"name": "master-color-mixer",
 			"title": "Color Mixer",
 			"data": "demos/src/palettes/master-palette.json",
-			"js": "demos/src/color-mixer/main.js",
-			"sass": "demos/src/color-mixer/main.scss",
+			"js": "demos/src/color-mixer/color-mixer.js",
+			"sass": "demos/src/color-mixer/color-mixer.scss",
 			"template": "demos/src/color-mixer/index.mustache",
 			"display_html": false,
 			"brands":	[
@@ -61,7 +61,7 @@
 			"title": "Primary Palette",
 			"data": "demos/src/palettes/master-palette.json",
 			"template": "demos/src/primary-palette.mustache",
-			"description": "color is a fundamental attribute of the FT visual language. Consistent use of the primary palette across the platform strengthens our brand by evoking familiarity and trust.",
+			"description": "Color is a fundamental attribute of the FT visual language. Consistent use of the primary palette across the platform strengthens our brand by evoking familiarity and trust.",
 			"display_html": false,
 			"brands": [
 				"master"
@@ -115,20 +115,20 @@
 			"name": "internal-contrast-checker",
 			"title": "Contrast ratio checker",
 			"data": "demos/src/palettes/internal-palette.json",
-			"js": "demos/src/sandboxes/contrast-checker.js",
-			"sass": "demos/src/sandboxes/contrast-checker.scss",
-			"template": "demos/src/sandboxes/contrast-checker.mustache",
+			"js": "demos/src/contrast-checker/contrast-checker.js",
+			"sass": "demos/src/contrast-checker/contrast-checker.scss",
+			"template": "demos/src/contrast-checker/index.mustache",
 			"display_html": false,
 			"brands": [
 				"internal"
 			]
 		},
 		{
-			"name": "internal-colors-mix",
+			"name": "internal-color-mixer",
 			"title": "Color Mixer",
 			"data": "demos/src/palettes/internal-palette.json",
-			"js": "demos/src/sandboxes/colors-mix.js",
-			"sass": "demos/src/color-mixer/main.scss",
+			"js": "demos/src/color-mixer/color-mixer.js",
+			"sass": "demos/src/color-mixer/color-mixer.scss",
 			"template": "demos/src/color-mixer/index.mustache",
 			"display_html": false,
 			"brands":	[
